Implement `mapping_context` manager
Removes \`globs.py\` and weird global stuff!
